+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased
 ----------
 - Added `cache_maps` property to `normalize::Normalizer` type
+- Introduced `normalize::Reason` enum to provide best guess at why normalization
+  was not successful as part of the `normalize::UserMeta::Unknown` variant
 - Reduced number of allocations performed on address normalization and
   process symbolization paths